third_party/WebKit/LayoutTests/svg/dom/SVGStaticListTearOff-contextElement-crash.html

<html>
<body>
<script src="../../resources/js-test.js"></script>
<script>
if (window.testRunner)
    testRunner.dumpAsText();

var mask = document.createElementNS("http://www.w3.org/2000/svg", "mask");
var maskObservation = internals.observeGC(mask);
var staticList = mask.requiredExtensions;
mask = null;
gc();

staticList.clear(); // This should not cause crash in ASAN.

// mask should be kept alive from "staticList"
shouldBeFalse('maskObservation.wasCollected');
</script>
This test passes if it doesn't crash in ASAN.
</body>
</html>
